admin/pyproject.toml
[tool.poetry]
name = "admin"
version = "1.0.0"
description = ""
authors = ["Ivan Nedobezhkin <inedobezkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-split-settings = "^1.2.0"
python-dotenv = "^1.0.0"
django-debug-toolbar = "^4.2.0"
uWSGI = "^2.0.22"
psycopg2-binary = "^2.9.7"
django-ckeditor = "^6.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
